refactor: clean up type-safe client side env vars
