+++ template/addons/next-auth/api-handler-prisma.ts
@@ -5,7 +5,7 @@ import CredentialsProvider from "next-auth/providers/credentials";
 // Prisma adapter for NextAuth, optional and can be removed
 import { PrismaAdapter } from "@next-auth/prisma-adapter";
 import { prisma } from "../../../server/db/client";
-import { env } from "../../../server/env.mjs";
+import { env } from "../../../env/server-env.mjs";
 
 export const authOptions: NextAuthOptions = {
   // Include user.id on session
