+++ template/base/next.config.mjs
@@ -1,5 +1,4 @@
-import { env as serverEnv } from "./src/server/env.mjs";
-import { env as clientEnv } from "./src/client/env.mjs"
+import { env } from './src/env/server-env.mjs'
 
 /**
  * Don't be scared of the generics here.
