--- template/base/src/server/env-schema.mjs
@@ -1,5 +1,0 @@
-import { z } from "zod";
-
-export const envSchema = z.object({
-  // Specify your environment variables schema here
-});
